requirements.txt
beautifulsoup4>=4.12,<5.0
httpx>=0.27,<1.0
openai>=1.13,<2.0
playwright>=1.45,<2.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
selenium>=4.22,<5.0
structlog>=24.1,<25.0
tenacity>=8.2,<9.0
typer>=0.12,<1.0
rich>=13.7,<14.0
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
jinja2>=3.1,<4.0
keyring>=24.3,<25.0
pytest>=7.4,<8.0

